requirements.txt
langchain
openai
python-dotenv
langchain-community
langchain-pinecone
langchain-openai
pinecone-client
unstructured
unstructured[pdf]
tiktoken
poppler-utils
# detectron2@git+https://github.com/facebookresearch/detectron2.git@v0.6#egg=detectron2
